chore: wip refactor decorators
