+++ packages/core/src/__tests__/config.ts
@@ -1,5 +1,5 @@
-import { DecoratorConfig, LintConfig, RuleConfig } from '../../config/config';
-import { defaultPlugin } from '../../config/builtIn';
+import { DecoratorConfig, LintConfig, RuleConfig } from '../config/config';
+import { defaultPlugin } from '../config/builtIn';
 
 export function makeConfig(rules: Record<string, RuleConfig>, decorators?:  Record<string, DecoratorConfig>) {
   return new LintConfig({
