+++ packages/core/src/config/builtIn.ts
@@ -4,6 +4,7 @@ import minimal from './minimal';
 import { LintRawConfig, Plugin } from './config';
 
 import * as builtinRules from '../rules/builtin';
+import * as builtinDecorators from '../decorators/builtin';
 
 export const builtInConfigs: Record<string, LintRawConfig> = {
   recommended,
@@ -18,6 +19,6 @@ export const defaultPlugin: Plugin = {
   id: '', // default plugin doesn't have id
   rules: builtinRules.rules,
   preprocessors: builtinRules.preprocessors,
-  decorators: builtinRules.decorators,
+  decorators: builtinDecorators.decorators,
   configs: builtInConfigs,
 }
