+++ packages/core/src/decorators/builtin.ts
@@ -1,0 +1,7 @@
+import * as oas3 from './oas3/index';
+import * as oas2 from './oas2/index';
+
+export const decorators = {
+  oas3: oas3.decorators,
+  oas2: oas2.decorators,
+};
