+++ packages/core/src/decorators/common/info-description-override.ts
