+++ packages/core/src/decorators/common/registry-dependencies.ts
