+++ packages/core/src/decorators/common/remove-x-internal.ts
