+++ packages/core/src/decorators/common/tag-description-override.ts
